+++ src/RoutingList/constants.js
@@ -1,0 +1,7 @@
+import { FormattedMessage } from 'react-intl';
+
+export const VISIBLE_COLUMNS = ['name'];
+
+export const COLUMN_MAPPING = {
+  name: <FormattedMessage id="ui-users.permissions.assignedUsers.name" />,
+};