+++ src/RoutingList/hooks/useRoutingListMutation/useRoutingListMutation.js
@@ -1,0 +1,39 @@
+import { useMutation } from 'react-query';
+
+import { useOkapiKy } from '@folio/stripes/core';
+
+import { ROUTING_LIST_API } from '../../../common/constants';
+
+export const useRoutingListMutation = () => {
+  const ky = useOkapiKy();
+
+  const createMutationFn = (values = {}) => {
+    return ky.post(ROUTING_LIST_API, {
+      json: values,
+    });
+  };
+
+  const updateMutationFn = (values = {}) => {
+    return ky.put(`${ROUTING_LIST_API}/${values.id}`, {
+      json: values,
+    });
+  };
+
+  const deleteMutationFn = (listId) => {
+    console.log('listId:', listId);
+    return ky.delete(`${ROUTING_LIST_API}/${listId}`).json();
+  };
+
+  const { mutateAsync: createListing, isCreating } = useMutation({ mutationFn: createMutationFn });
+  const { mutateAsync: updateListing, isLoading: isUpdating } = useMutation({ mutationFn: updateMutationFn });
+  const { mutateAsync: deleteListing, isLoading: isDeleting } = useMutation({ mutationFn: deleteMutationFn });
+
+  return {
+    createListing,
+    deleteListing,
+    updateListing,
+    isCreating,
+    isDeleting,
+    isUpdating,
+  };
+};